==> 1.Preprocesamiento-variables-dummy(One-Hot Encoding).ipynb <==
# Preprocesamiento de Variables Dummy

Existen muchos métodos de machine learning para los cuales es necesario convertir las variables categóricas en 1’s y 0’s. 
La manera más sencilla de hacerlo es crear variables dummy (también llamadas ficticias), mediante un proceso conocido como one-hot encoding, el cual crea variables binarias, representando en cada fila la existencia de la característica con un “1” o su ausencia con un “0”. 

import pandas as pd

hidroponia=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Variables Dummy (One-Hot Encoding)/Hidroponia_original.csv', encoding='cp1252')

hidroponia.info()

hidroponia

Hortaliza_dummy = pd.get_dummies(hidroponia['Hortaliza'], prefix='Hortaliza')

Sistema_dummy = pd.get_dummies(hidroponia['Sistema'], prefix='Sistema')

hidroponia = pd.concat([hidroponia, Hortaliza_dummy], axis = 1)

hidroponia = pd.concat([hidroponia, Sistema_dummy], axis = 1)

hidroponia

hidroponia_dummies = hidroponia.drop (['pH', 'EC', 'Hortaliza', 'Sistema', 'Duracion_Ciclo '], axis =1) 

hidroponia_dummies

hidroponia_dummies_columnas = hidroponia.drop (hidroponia.columns[[0, 1, 2, 3, 4]], axis =1) 

hidroponia_dummies_columnas

hidroponia_dummies_iloc=hidroponia.drop (hidroponia.iloc[:, 0:5], axis=1)

hidroponia_dummies_iloc

hidroponia_dummies.to_csv('C:/Users/pc/Archivos en Jupyter/Variables Dummy (One-Hot Encoding)/Hidroponia_original.csv')
