==> 3-ML_Feature-Selection.ipynb <==
## Machine Learning Model Building Pipeline: Feature Selection



1. Data Analysis
2. Feature Engineering
3. Feature Selection
4. Model Building

**This is the notebook for step 3: Feature Selection**

## House Prices dataset: Feature Selection

En las siguientes celdas, seleccionaremos un grupo de variables, las más predictivas, para construir nuestro modelo de aprendizaje automático.

### ¿Por qué seleccionamos variables?

- Para producción: menos variables significan requisitos de entrada de clientes más pequeños (por ejemplo, clientes que completan un formulario en un sitio web o aplicación móvil) y, por lo tanto, menos código para el manejo de errores. Esto reduce las posibilidades de introducir errores.

- Para el rendimiento del modelo: Menos variables significan modelos más simples, más interpretables y de mejor generalización


**Seleccionaremos las variables mediante la regresión de Lasso: Lasso tiene la propiedad de poner a cero el coeficiente de las variables no informativas. De esta manera podemos identificar esas variables y eliminarlas de nuestro modelo final.**


### Estableciendo la semilla

Es importante tener en cuenta que estamos diseñando variables y preprocesando datos con la idea de implementar el modelo. Por lo tanto, de ahora en adelante, para cada paso que incluya algún elemento de aleatoriedad, es extremadamente importante que **establezcamos la semilla**. De esta manera, podemos obtener reproducibilidad entre nuestra investigación y nuestro código de desarrollo.

**Siempre establece las semillas**.

# to handle datasets
import pandas as pd
import numpy as np

# for plotting
import matplotlib.pyplot as plt

# to build the models
from sklearn.linear_model import Lasso
from sklearn.feature_selection import SelectFromModel

# to visualise al the columns in the dataframe
pd.pandas.set_option('display.max_columns', None)

# load the train and test set with the engineered variables

X_train = pd.read_csv('xtrain.csv')
X_test = pd.read_csv('xtest.csv')

X_train.head()
#X_test.head()

# capturar el objetivo target (recuerde que el objetivo se transforma en registro)
y_train = X_train['SalePrice']
y_test = X_test['SalePrice']

# elimine variables innecesarias de nuestros conjuntos de entrenamiento y prueba
X_train.drop(['Id', 'SalePrice'], axis=1, inplace=True)
X_test.drop(['Id', 'SalePrice'], axis=1, inplace=True)

### Feature Selection

Avancemos y seleccionemos un subconjunto de las funciones más predictivas. Hay un elemento de aleatoriedad en la regresión de Lasso, así que recuerda establecer la semilla.

# Haremos el ajuste del modelo y la selección de características.
# en total en unas pocas líneas de código

# primero, especificamos el modelo de regresión de Lasso, y
# seleccionar un alfa adecuado (equivalente a penalización).
# Cuanto mayor sea el alfa, menos funciones se seleccionarán.

# Luego usamos el objeto selectFromModel de sklearn, que
# seleccionará automáticamente las características cuyos coeficientes no sean cero

# recuerda configurar la semilla, el estado aleatorio en esta función
sel_ = SelectFromModel(Lasso(alpha=0.005, random_state=0))

# train Lasso model and select features
sel_.fit(X_train, y_train)

# vamos a visualizar aquellas características que fueron seleccionadas.
# (características seleccionadas marcadas con Verdadero)

sel_.get_support()

# imprimamos el número de características totales y seleccionadas

# así es como podemos hacer una lista de las características seleccionadas
selected_feats = X_train.columns[(sel_.get_support())]

# imprimamos algunas estadísticas
print('total features: {}'.format((X_train.shape[1])))
print('selected features: {}'.format(len(selected_feats)))
print('features with coefficients shrank to zero: {}'.format(
    np.sum(sel_.estimator_.coef_ == 0)))

# imprimir las características seleccionadas
selected_feats

### Identificamos las variables seleccionadas

# esta es una forma alternativa de identificar las características seleccionadas
# basado en los coeficientes de regularización distintos de cero:

selected_feats = X_train.columns[(sel_.estimator_.coef_ != 0).ravel().tolist()]

selected_feats

pd.Series(selected_feats).to_csv('selected_features.csv', index=False)
