==> 3. Preprocesamiento-valores perdidosy filas duplicadas.ipynb <==
# Pre-Procesamiento de Datos con Python: Valores Perdidos y Filas Duplicadas

Valores Perdidos y Filas Duplicadas en Python y la manera de detectar los valores perdidos dentro de un archivo y cómo podías borrar las filas o las columnas en las que éstos se encontraban. 

Algunas técnicas que se pueden aplicar para calcular los valores perdidos, sin embargo, debo aclarar que el reemplazo de valores es un tema muy delicado, ya que, si la técnica que se utiliza no es la adecuada, los valores creados en lugar de ayudar en el análisis de datos pueden afectar negativamente el estudio que se esté realizando. 

import pandas as pd

importar datos

niveles=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Pre-Procesamiento de Datos con Python Valores Perdidos y Filas Duplicadas/Niveles de contaminacion.csv', engine ='python')

niveles.info()

niveles.head()

niveles['PM10'].isnull()

niveles.isnull().sum()

compras=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Pre-Procesamiento de Datos con Python Valores Perdidos y Filas Duplicadas/Patron de Compra con NAs.csv', engine ='python')

compras.info()

compras

compras2=compras.dropna()

compras2

compras3=compras.dropna(axis=1)

compras3

compras4=compras.dropna(axis=1, thresh=12)

compras4

niveles.drop_duplicates()

niveles.dropna(thresh=8)

niveles.mean()

niveles.std()
