==> 4.Preprocesamiento-reemplazo-valores-perdidos.ipynb <==
# Reemplazando los Valores Perdidos

Se ven algunas técnicas que se pueden aplicar para calcular los valores perdidos, sin embargo, debo aclarar que el reemplazo de valores es un tema complicado, ya que, si la técnica que se utiliza no es la adecuada, los valores creados en lugar de ayudar en el análisis de datos pueden afectar negativamente el estudio que se esté realizando. Debido a que depende de muchos factores la decisión acerca de qué método de sustitución de valores faltantes es mejor aplicar, lo que haremos será ver únicamente los métodos más simples de calcular, pero que no necesariamente son los indicados en todas las situaciones. 

import numpy as np
import pandas as pd

niveles=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Pre-Procesamiento de Datos en Python Reemplazando los Valores Perdidos/Niveles de contaminacion.csv', engine ='python')

niveles. info()

niveles.head()

niveles

niveles1=niveles

niveles1. fillna(0)

niveles2 = niveles.apply(lambda x: x.fillna(x.mean()), axis=0)

niveles2

niveles3 = niveles.apply(lambda x: x.fillna(x.median()), axis=0)
niveles3

compras=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Pre-Procesamiento de Datos en Python Reemplazando los Valores Perdidos/Patron de Compra con NAs.csv', engine ='python', index_col=0)

compras.info()

compras

compras1=compras("Profesion").fillna("Desconocida", inplace=True)

compras2=compras.fillna(method='ffill')
compras2

compras3=compras.fillna(method='bfill')
compras3
