==> 5-ML_Resumen-Model-Deployment.ipynb <==
## Machine Learning Pipeline: Resumen del desarrollo


En los cuadernos anteriores, trabajamos con los pasos típicos de canalización de Machine Learning para construir un modelo de regresión que nos permita predecir los precios de la vivienda. Brevemente, transformamos las variables en el conjunto de datos para que sean adecuadas para su uso en un modelo de regresión, luego seleccionamos las variables más predictivas y finalmente entrenamos nuestro modelo.

Ahora, queremos implementar nuestro modelo. Queremos crear una API, a la que podamos llamar con nuevos datos, con nuevas características sobre las casas, para obtener una estimación del precio de venta. Para hacerlo, necesitamos escribir código de una manera muy específica. Le mostraremos cómo escribir código de producción en las siguientes secciones.

Aquí, resumiremos las piezas clave de código que necesitamos llevar adelante para este proyecto en particular, para poner nuestro modelo en producción.

Sigamos adelante y comencemos.

### Establecer la semilla

Es importante tener en cuenta que estamos diseñando variables y preprocesando datos con la idea de implementar el modelo. Por lo tanto, de ahora en adelante, para cada paso que incluya algún elemento de aleatoriedad, es extremadamente importante que **establezcamos la semilla**. De esta manera, podemos obtener reproducibilidad entre nuestra investigación y nuestro código de desarrollo.

Esta es quizás una de las lecciones más importantes que debes aprender de este curso: **Siempre establece las semillas**.

Avancemos y carguemos el conjunto de datos.

# to handle datasets
import pandas as pd
import numpy as np

# to divide train and test set
from sklearn.model_selection import train_test_split

# feature scaling
from sklearn.preprocessing import MinMaxScaler

# to build the models
from sklearn.linear_model import Lasso

# to evaluate the models
from sklearn.metrics import mean_squared_error, r2_score
from math import sqrt

# to persist the model and the scaler
import joblib

# to visualise al the columns in the dataframe
pd.pandas.set_option('display.max_columns', None)

import warnings
warnings.simplefilter(action='ignore')

## Load data

Necesitamos los datos de entrenamiento para entrenar nuestro modelo en el entorno de producción. 

# load dataset
data = pd.read_csv('houseprice.csv')
print(data.shape)
data.head()

## Separamos el dataset en train and test

X_train, X_test, y_train, y_test = train_test_split(
    data,
    data['SalePrice'],
    test_size=0.1,
    # we are setting the seed here
    random_state=0)

X_train.shape, X_test.shape

X_train.head()

## Selecccionamos las features

# load selected features
features = pd.read_csv('selected_features.csv')

# Added the extra feature, LotFrontage
features = features['0'].to_list() + ['LotFrontage']

print('Number of features: ', len(features))

## Ingenieria de los valores missing

### Variables categoricas

Para las variables categóricas, reemplazaremos los valores faltantes con la cadena "missing".

# hacer una lista de las variables categóricas que contienen valores faltantes

vars_with_na = [
    var for var in features
    if X_train[var].isnull().sum() > 0 and X_train[var].dtypes == 'O'
]

# mostrar variables categóricas que diseñaremos:
vars_with_na

Tenga en cuenta que tenemos muchas menos variables categóricas con valores faltantes que en nuestro conjunto de datos original. Pero aún usamos variables categóricas con NA para el modelo final, por lo que debemos incluir esta parte de la lógica de ingeniería de características en la implementación de la pipeline.


X_train[vars_with_na] = X_train[vars_with_na].fillna('Missing')
X_test[vars_with_na] = X_test[vars_with_na].fillna('Missing')

# chequeamos que no hay missing informationen las variables
X_train[vars_with_na].isnull().sum()

### Variables numéricas

Para diseñar valores faltantes en variables numéricas, haremos lo siguiente:

- agregar una variable indicadora de valor faltante binario
- y luego reemplace los valores que faltan en la variable original con el modo


# hacer una lista de las variables numéricas que contienen valores faltantes:

vars_with_na = [
    var for var in features
    if X_train[var].isnull().sum() > 0 and X_train[var].dtypes != 'O'
]

#  Mostrar variables numéricas con NA
vars_with_na


var = 'LotFrontage'

# calculate the mode
mode_val = X_train[var].mode()[0]
print('mode of LotFrontage: {}'.format(mode_val))

# replace missing values by the mode
# (in train and test)
X_train[var] = X_train[var].fillna(mode_val)
X_test[var] = X_test[var].fillna(mode_val)

## Variables temporales

Una de nuestras variables temporales fue seleccionada para ser utilizada en el modelo final: 'YearRemodAdd'

Así que necesitamos implementar el bit de código que lo crea..

# creamos la variable tempporal "elapsed years"


def elapsed_years(df, var):
    # obtenermos la diferencia entre la variable 
    # con el año en el que la casa ha sido vendida
    
    df[var] = df['YrSold'] - df[var]
    
    return df

X_train = elapsed_years(X_train, 'YearRemodAdd')
X_test = elapsed_years(X_test, 'YearRemodAdd')

### Transformacion de la variable numerica

# aplicamos la función logarítmica a las variables que
# fueron seleccionados (y el objetivo):

for var in ['LotFrontage', '1stFlrSF', 'GrLivArea', 'SalePrice']:
    X_train[var] = np.log(X_train[var])
    X_test[var] = np.log(X_test[var])

## Variables categoricas

### Agrupar etiquetas raras


# capturemos primero las variables categóricas
cat_vars = [var for var in features if X_train[var].dtype == 'O']

cat_vars



def find_frequent_labels(df, var, rare_perc):
    
    # función encuentra las etiquetas que son compartidas por más de
    # un cierto % de las casas en el conjunto de datos

    df = df.copy()

    tmp = df.groupby(var)['SalePrice'].count() / len(df)

    return tmp[tmp > rare_perc].index

for var in cat_vars:
    
    # encontrar las categorías frecuentes
    frequent_ls = find_frequent_labels(X_train, var, 0.01)
    print(var)
    print(frequent_ls)
    print()
    
    # reemplace las categorías raras por la cadena "Raro"
    X_train[var] = np.where(X_train[var].isin(
        frequent_ls), X_train[var], 'Rare')
    
    X_test[var] = np.where(X_test[var].isin(
        frequent_ls), X_test[var], 'Rare')

### Codificación de variables categóricas



# esta función asignará valores discretos a las cadenas de las variables,
# para que el valor menor corresponda a la categoría que muestra el menor
# precio medio de venta de la casa

def replace_categories(train, test, var, target):

    # ordena las categorías en una variable a partir de la que tiene la menor
    # precio de venta de la casa, al de mayor
    ordered_labels = train.groupby([var])[target].mean().sort_values().index
    
    # crear un diccionario de categorías ordenadas a valores enteros
   
    ordinal_label = {k: i for i, k in enumerate(ordered_labels, 0)}
    
    # usar el diccionario para reemplazar las cadenas categóricas por números enteros
  
    train[var] = train[var].map(ordinal_label)
    test[var] = test[var].map(ordinal_label)
    
    print(var)
    print(ordinal_label)
    print()

for var in cat_vars:
    replace_categories(X_train, X_test, var, 'SalePrice')

# check absence of na
[var for var in features if X_train[var].isnull().sum() > 0]

# check absence of na
[var for var in features if X_test[var].isnull().sum() > 0]

### Escalado de características

Para su uso en modelos lineales, las características deben escalarse o normalizarse. En la siguiente sección, escalaré las funciones entre los valores mínimo y máximo:

# capturamos el target
y_train = X_train['SalePrice']
y_test = X_test['SalePrice']

# set up scaler
scaler = MinMaxScaler()

# train scaler
scaler.fit(X_train[features])

# exploramos maximo valor para las variables
scaler.data_max_

# explore minimo valor para las variables
scaler.data_min_


# transformamos el set de train y test y agregamos las variables Id y SalePrice
X_train = scaler.transform(X_train[features])
X_test = scaler.transform(X_test[features])

## Entrenamiento la regresion lineal: Lasso

# set up the model
# remember to set the random_state / seed

lin_model = Lasso(alpha=0.005, random_state=0)

# train the model
lin_model.fit(X_train, y_train)

# we persist the model for future use
joblib.dump(lin_model, 'lasso_regression.pkl')

# evaluar el modelo:
# ====================

# recuerda que registramos transformado la salida (SalePrice)
# Para obtener el verdadero rendimiento del Lasso
# necesitamos transformar tanto el objetivo como las predicciones
# volver a los valores originales de los precios de la vivienda.

# Evaluaremos el desempeño utilizando el error cuadrático medio y
# la raíz del error cuadrático medio y r2

# hacer predicciones para el juego de trenes
pred = lin_model.predict(X_train)

# determine mse and rmse
print('train mse: {}'.format(int(
    mean_squared_error(np.exp(y_train), np.exp(pred)))))
print('train rmse: {}'.format(int(
    sqrt(mean_squared_error(np.exp(y_train), np.exp(pred))))))
print('train r2: {}'.format(
    r2_score(np.exp(y_train), np.exp(pred))))
print()

# make predictions for test set
pred = lin_model.predict(X_test)

# determine mse and rmse
print('test mse: {}'.format(int(
    mean_squared_error(np.exp(y_test), np.exp(pred)))))
print('test rmse: {}'.format(int(
    sqrt(mean_squared_error(np.exp(y_test), np.exp(pred))))))
print('test r2: {}'.format(
    r2_score(np.exp(y_test), np.exp(pred))))
print()

print('Average house price: ', int(np.exp(y_train).median()))
