==> 5. Filtrando Información.ipynb <==
# Filtrando Información en Python

import pandas as pd

atletas=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Filtrando Información en Python/Corredores.csv', index_col=0)

atletas.info()

atletas.head()

atletas.iloc[0:15]

atletas.iloc[:, 0:3]

atletas.iloc[:, [0, 3, 4]]

atletas.iloc[[0, 5, 7], [0, 3, 4]]

atletas.loc[["MEX","ARG"],["Genero", "Tiempo"]]

atletas.loc[["MEX"],"Genero": "Tiempo"]

atletas_Mujeres=atletas[atletas["Genero"]=="Female"]
atletas_Mujeres

atletas_Edad=atletas[atletas["Edad"] >40]
atletas_Edad
