==> 6. Categorización de variables.ipynb <==
# Categorización de Variables en Python

En esta sección explicamos cómo podemos dividir una información numérica en diferentes categorías. 
Esta conversión es útil cuando no nos interesa saber la cantidad numérica que un individuo poseee en una de sus características, sino a qué nivel pertenece de acuerdo a esa cantidad. 
Por ejemplo, tal vez no nos interesa saber exactamente cuanto mide una persona, sino simplemente saber si ésta es alta, bajita o de estatura regular. 
Lo que se hace en estos casos es determinar cuales categorías queremos formar y cuales son los valores numéricos que queremos que represente cada una de ellas

import pandas as pd
import math
import numpy as np

atletas=pd.read_csv ('C:/Users/pc/Archivos en Jupyter/Filtrando Información en Python/Corredores.csv', index_col=0)

atletas.info()

atletas.head()

valores = [100, 200, 300, 400]
valores

categorias= ["Valor", "Regular", "Lento"]

atletas["Velocidad"]=pd.cut(x=atletas["Tiempo"], bins= [100,200,300])

atletas
