==> 14.Reemplazando valores perdidos con K-NN.ipynb <==
## Reemplazando valores perdidos con K-NN

Nota: si quisieramos utilizar este algoritmo con variables categoricas, primero debemos convertirlas a numéricas

#### 1. Carga de librerias y datos

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
from matplotlib import rcParams

from sklearn.impute import KNNImputer

niveles=pd.read_csv('C:/Users/pc/Desktop/rociochavez/Rocio-Chavez-youtube-Files-master/Niveles de contaminacion.csv')

niveles

#### 2. Reemplazo utilizando solo un vecino

#creamos el modelo para llevar adelante la imputación

imputacion = KNNImputer (n_neighbors=1)

imputacion.n_neighbors

imputacion.metric

#### 3. Aplicamos el modelo de imputación al objeto niveles

niveles_knn=imputacion.fit_transform(niveles)

niveles_knn

##### 4. Convertimos el array en un dataframe

niveles_knn=pd.DataFrame(niveles_knn)
niveles_knn

#### Le asignamos los nombres a las columnas del data frame que creamos

niveles_knn.columns= list(niveles.columns.values)
niveles_knn

#### 2. Reemplazo utilizando solo un vecino

imputacion2 = KNNImputer (n_neighbors=2)

niveles_knn2=imputacion2.fit_transform(niveles)
niveles_knn2=pd.DataFrame(niveles_knn2)
niveles_knn2.columns= list(niveles.columns.values)
niveles_knn2
