==> 15.Eliminación de variables con varianza igual a cero.ipynb <==
## Eliminación de variables con varianza igual a cero

#### 1. Carga de librerias y datos

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
from matplotlib import rcParams

datos =pd.read_csv('C:/Users/pc/Desktop/rociochavez/Rocio-Chavez-youtube-Files-master/Ejemplo varianza.csv', index_col=0)

datos

#### 2. Obtenemos la varianza de la columna A

varianza_A=datos['A'].var()

varianza_A

#### 3. Utilizamos una lista y un ciclo for para detectar todas las columnas que tengan varianza igual a cero

cant_col=datos.shape[1]
cant_col

col_var0=[]

for i in range(0, cant_col):
    varianza=datos.iloc[:,i].var()
    print(datos.columns.values[i], varianza)
    if varianza==0:
        col_var0.append(i)

col_var0

#### 4 Borramos las columnas cuya varianza sea igual a cero

datos.drop(datos.columns.values[col_var0], axis=1)
datos
