==> 16.Transformacion de variables Categoricas a numericas en Python.ipynb <==
## Transformacion de variables Categoricas a numericas en Python

#-variables ordinales: niveles de educación. Sus valores indican el nivel de presencia, o ausencia, de un aspecto concreto
#-Variable nominal: genero. nominales indican una categoría


#### 1. Carga de librerias y datos

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
from matplotlib import rcParams
from sklearn.preprocessing import OneHotEncoder
from sklearn.compose import ColumnTransformer

clientes=pd.read_csv('C:/Users/pc/Desktop/rociochavez/Rocio-Chavez-youtube-Files-master/Clientes Creado3.csv', index_col=0)

clientes

clientes.info()

### 1. Si todas las variables categoricas fueran nominales

clientes_categ=clientes.select_dtypes(include=['object']).columns
clientes_categ

#### 2. Indicamos la transformación que queremos que se lleve a cabo

#Creamos variables dummy

t=[('nominales', OneHotEncoder(drop='if_binary'), clientes_categ)]

#### 3. Aplicamos las transformaciones

col_transform=ColumnTransformer(transformers=t)
col_transform

clientes_numericas=col_transform.fit_transform(clientes)
clientes_numericas

clientes_nombres=col_transform.get_feature_names_out()
clientes_nombres

col_transform=ColumnTransformer(transformers=t, remainder='passthrough')

clientes_numericas=col_transform.fit_transform(clientes)
clientes_numericas

clientes_numericas_df=pd.DataFrame(clientes_numericas)
clientes_numericas_df

clientes_numericas_df.columns=['Mujer', '1 a 3 Hijos', '3 o mas hijos', 'Ningun hijo', 'Licenciatura', 'Maestria', 'Preparatoria','Ingreso', 'edad']
clientes_numericas_df


### 2. Si tenemos variables nominales y Ordinales

#Tomamos las variables genero y estudios como nominales y las variables hijos como ordinal

tr_nom=ColumnTransformer([('nominales', OneHotEncoder(drop='if_binary'), ['Genero', 'Estudios'])], remainder='passthrough')

clientes_numericas2=tr_nom.fit_transform(clientes)
clientes_numericas2

tr_nom.get_feature_names_out()

clientes_numericas_df2=pd.DataFrame(clientes_numericas2)
clientes_numericas_df2

#### 2.1 Creamos los codificadores de las variables ordinales con un orden menor a mayor importancia

from sklearn.preprocessing import OrdinalEncoder

tr_ord=ColumnTransformer([('hijos', OrdinalEncoder(categories=[('Mas de 3', '1 a 3', 'Ninguno')]), [4])],remainder='passthrough')

clientes_numericas2=tr_ord.fit_transform(clientes_numericas_df2)
clientes_numericas2

#Ejemplo de codigo si tuvieras mas de una variable ordinal


from sklearn.preprocessing import OrdinalEncoder

tr_ord2=ColumnTransformer([('hijos',  OrdinalEncoder(categories=[('Mas de 3', '1 a 3', 'Ninguno')]), ['Hijos']),
                          ('estudios', OrdinalEncoder(categories=[('Preparatoria', 'Licenciatura', 'Maestria')]), ['Estudios'])],
                          remainder='passthrough')
clientes_numericas4=tr_ord2.fit_transform(clientes)
clientes_numericas4

clientes_numericas_df3=clientes_numericas_df3[[1,2,3,4,0,5,6]]
clientes_numericas_df3



clientes_numericas_df3=pd.DataFrame(clientes_numericas2)
clientes_numericas_df3

clientes_numericas_df3.columns=['Mujer', 'Licenciatura', 'Maestria', 'Preparatoria','Hijos', 'Ingreso', 'Edad']

clientes_numericas_df3

#Guardamos la información en un csv

#clientes_numericas_df3.to_csv(ruta)
