==> 17.Creación de variables nuevas a partir de las ya existentes.ipynb <==
## Creación de variables nuevas a partir de las ya existentes

#### 1. Carga de librerias y datos

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
from matplotlib import rcParams
from sklearn.preprocessing import OneHotEncoder
from sklearn.compose import ColumnTransformer

hidroponia= pd.read_csv('C:/Users/pc/Desktop/rociochavez/Rocio-Chavez-youtube-Files-master/Hidroponia_original.csv', index_col=0, encoding='latin-1')

hidroponia

hidroponia['Hort Sist']=hidroponia['Hortaliza'] + '-' + hidroponia['Sistema']
hidroponia

#### 2. Conversion de variables Ordinales a numerica

#Creamos un diccionario con los valores que queremos asignar

diccionario={'Aeroponia': 1, 'Goteo': 2, 'NFT': 3}
hidroponia['Sistema Costo']=hidroponia['Sistema'].replace(diccionario)
hidroponia

#Eliminamos las columnas que ya no necesitamos
hidroponia2 = hidroponia.drop(['Hortaliza', 'Sistema'], axis=1)
hidroponia2

#### 3. Multiplicacion de los valores de dos o mas variables existentes

import pandas as pd

df = pd.DataFrame({'A': ['foo', 'bar', 'baz'], 'B': ['foobar', 'barbaz', 'quxqux']})

# Check if the substring 'bar' is present in column 'B'
df['B_has_substring'] = df['B'].str.contains('bar')

print(df)

hidroponia2.info()

#creamos una variable llamada duración para detectar los cultivos que podrían durar 100 dias o mas
hidroponia2['Hort Sist2']=hidroponia2['Hort Sist'].str.contains('Flor-NFT')
hidroponia2

for i in range(hidroponia2.shape[0]):
    if hidroponia2.iloc[i, 6]:
        hidroponia2.iloc[i,6]=2
    else:
        hidroponia2.iloc[i,6]=1
hidroponia2
