==> 19. SVD Aplicado a Sistemas de Recomendacion.ipynb <==
# Recommendation Engine / System

import pandas as pd
import numpy as np

productos = pd.read_csv("Peliculas.csv", sep = ';', index_col=0)
productos

from scipy.linalg import svd

#calcula los componnentes principales usando svd

U, s, V_transp=svd(productos, full_matrices=False)

#matriz de vectores singulares
U

#matriz de observaciones
V_transp

#Mostrar los valores singulare
s

sigma_s=[]
for i in range(len(s)):
    a=(s[i]/(sum(s)))*100
    sigma_s.append(a)

sigma_s

#Le indicamos la cantidad de valores singulares a tomar en cuenta
k=3

#Matriz de elementos
matriz_usuarios=U[:, :k]

matriz_usuarios_df=pd.DataFrame(np.round(matriz_usuarios,2))
matriz_usuarios_df

#Matriz de Elementos
matriz_elementos=V_transp[:k, :]
matriz_elementos

matriz_elementos_df=pd.DataFrame(np.round(matriz_elementos,2))
matriz_elementos_df

#calculamos los valores predichos de preferencia de los usuarios existentes para poderles recomendar mas peliculas

ratings_predichos=np.dot(matriz_usuarios,matriz_elementos)
ratings_predichos

ratings_predichos_df=pd.DataFrame(np.round(ratings_predichos,2),columns=productos.columns)
ratings_predichos_df

#Recomendamos para un usuario especifico

userID=1

prediccciones_usuario=ratings_predichos_df.loc[userID]
prediccciones_usuario
